[trainer] move secondary methods into a separate file (#10363)
* move secondary methods into a separate file

* cleanup

* style
